+++ adelie/__init__.py
@@ -12,4 +12,4 @@ def _get_version() -> str:
     except Exception:
         pass
     return "0.0.0"
-__version__ = "0.3.2"
+__version__ = "0.3.3"